UAT_TabelBall/backend/requirements.txt
APScheduler==3.10.4
attrs==23.2.0
Babel==2.15.0
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
cachelib==0.9.0
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF==1.2.1
future==1.0.0
fuzzywuzzy==0.18.0
greenlet==3.1.1
h11==0.14.0
idna==3.7
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
Levenshtein==0.26.0
lxml==5.3.0
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
numpy==2.0.0
openpyxl==3.1.5
outcome==1.3.0.post0
pandas==2.2.2
platformdirs==4.3.6
PyMySQL==1.1.1
pyodbc==5.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-Levenshtein==0.26.0
python-socketio==5.11.3
pytz==2024.1
pyzk==0.9
RapidFuzz==3.10.0
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
selenium==4.22.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.35
suds-py3==1.4.5.0
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.0
zeep==4.3.1
